- Fixed item picture not displayed in map editor selection area for those items beyond the old limit range (by Buggler)
git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@5883 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Editor/EditorItems.cpp
@@ -229,12 +229,12 @@ void InitEditorItemsInfo(UINT32 uiItemType)
 	HVOBJECT hVObject;
 	UINT32 uiVideoObjectIndex;
 	UINT16 usUselessWidth, usUselessHeight;
-	INT16 sWidth, sOffset, sStart;
+	INT32 sWidth, sOffset, sStart;
 	INT32 i, x, y;
 	UINT16 usCounter;
 	CHAR16 pStr[ 100 ];//, pStr2[ 100 ];
 	CHAR16 pItemName[SIZE_ITEM_NAME];
-	UINT8						ubBitDepth;
+	UINT8 ubBitDepth;
 	BOOLEAN fTypeMatch;
 	INT32 iEquipCount = 0;
